requirements.txt

Django==4.1.2
psycopg2-binary==2.8.6
django-randomslugfield==0.3.0

